Reformed style into a more advanced SASS
Exploded all files with the use of SASS import
Used SASS Selection Statements and used mixins to switch color
The code for the above statement is already done but it still needs
to be tested. The script is yet to follow. For this commit, it is
still focused on the CSS.